Play services update validation for app Splash Screen
Welcome to the Google-Play-Services-Update-Handling wiki!

It's often found that the google play services not updated in android
device and you are creating app using latest google play services which
cause error while you request for the location here is a class which
you can use to check whether google play services are updated or not
and open update screen and hence you can restrict the app opening
without updated google play services.

It is helpful to those whose are migrating their google location
services to SettingApi